Update fuzzer.cpp
The code doesn't account for command-line argument "-crash_retry 0". 
If provided, this argument will cause a "Run-Time Check Failure #3 - The variable 'result' is being used without being initialized."
This is an uncommon usecase, but I have encountered this while fuzzing a binary that keeps a handle to the crashing file for a bit too long. 
Using "-crash_retry 0" allows for fuzzing the binary, albeit flaky fuzzing.